requirements.txt
opencv-python>=4.1.1.26
lxml
tqdm
numpy
tensorflow>=2.3.0rc0
absl-py
easydict
matplotlib
pillow
python-multipart
Shapely